+++ api/typescript/src/engine/lib/kurtosis_context/grpc_node_engine_client.ts
@@ -1,6 +1,12 @@
-import {err, ok, Result} from "neverthrow";
-import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
-import type {ServiceError} from "@grpc/grpc-js";
+import * as jspb from 'google-protobuf';
+import * as google_protobuf_empty_pb from 'google-protobuf/google/protobuf/empty_pb';
+import { err, ok, Result } from 'neverthrow';
+import { Readable } from 'stream';
+
+import { ServiceGUID } from '../../../core/lib/services/service';
+import { LogLine } from '../../kurtosis_engine_rpc_api_bindings/engine_service_pb';
+import { NO_ERROR_ENCOUNTERED_BUT_RESPONSE_FALSY_MSG } from '../consts';
+
 import type {EngineServiceClient as EngineServiceClientNode} from "../../kurtosis_engine_rpc_api_bindings/engine_service_grpc_pb";
 import type {GenericEngineClient} from "./generic_engine_client";
 import type {
@@ -15,12 +21,20 @@ import type {
     GetUserServiceLogsResponse,
     StopEnclaveArgs
 } from "../../kurtosis_engine_rpc_api_bindings/engine_service_pb";
-import {NO_ERROR_ENCOUNTERED_BUT_RESPONSE_FALSY_MSG} from "../consts";
+import type {ClientReadableStream, ServiceError} from "@grpc/grpc-js";
+
+const GRPC_STREAM_RESPONSE_DATA_EVENT_NAME = 'data'
+const GRPC_STREAM_RESPONSE_ERROR_EVENT_NAME = 'error'
+const GRPC_STREAM_RESPONSE_END_EVENT_NAME = 'end'
+
+const USER_SERVICE_LOGS_READABLE_ERROR_EVENT_NAME = 'error'
+const USER_SERVICE_LOGS_READABLE_END_EVENT_NAME = 'end'
+const USER_SERVICE_LOGS_READABLE_CLOSE_EVENT_NAME = 'close'
 
 export class GrpcNodeEngineClient implements GenericEngineClient {
     private readonly client: EngineServiceClientNode
 
-    constructor(client: EngineServiceClientNode){
+    constructor(client: EngineServiceClientNode) {
         this.client = client
     }
 
@@ -39,7 +53,7 @@ export class GrpcNodeEngineClient implements GenericEngineClient {
                     }
                 } else {
 
-                    if(error.code === grpc_node.status.UNAVAILABLE){
+                    if (error.code === grpc_node.status.UNAVAILABLE) {
                         resolve(err(new Error("The Kurtosis Engine Server is unavailable and is probably not running; you " +
                             "will need to start it using the Kurtosis CLI before you can create a connection to it")));
                     }
@@ -119,7 +133,7 @@ export class GrpcNodeEngineClient implements GenericEngineClient {
         return ok(null);
     }
 
-    public async clean(cleanArgs: CleanArgs): Promise<Result<CleanResponse, Error>>{
+    public async clean(cleanArgs: CleanArgs): Promise<Result<CleanResponse, Error>> {
         const cleanPromise: Promise<Result<CleanResponse, Error>> = new Promise((resolve, _unusedReject) => {
             this.client.clean(cleanArgs, (error: ServiceError | null, response?: CleanResponse) => {
                 if (error === null) {
@@ -143,7 +157,7 @@ export class GrpcNodeEngineClient implements GenericEngineClient {
         return ok(cleanResponse);
     }
 
-    public async getEnclavesResponse(): Promise<Result<GetEnclavesResponse, Error>>{
+    public async getEnclavesResponse(): Promise<Result<GetEnclavesResponse, Error>> {
         const emptyArg: google_protobuf_empty_pb.Empty = new google_protobuf_empty_pb.Empty()
         const getEnclavesPromise: Promise<Result<GetEnclavesResponse, Error>> = new Promise((resolve, _unusedReject) => {
             this.client.getEnclaves(emptyArg, (error: ServiceError | null, response?: GetEnclavesResponse) => {
@@ -158,7 +172,7 @@ export class GrpcNodeEngineClient implements GenericEngineClient {
                 }
             })
         });
-        
+
         const getEnclavesResponseResult: Result<GetEnclavesResponse, Error> = await getEnclavesPromise;
         if (getEnclavesResponseResult.isErr()) {
             return err(getEnclavesResponseResult.error);
@@ -189,4 +203,69 @@ export class GrpcNodeEngineClient implements GenericEngineClient {
 
         return ok(getUserServiceLogsResponseResult.value);
     }
+
+    public async streamUserServiceLogs(getUserServiceLogsArgs: GetUserServiceLogsArgs): Promise<Result<Readable, Error>> {
+
+        const streamUserServiceLogsPromise: Promise<Result<ClientReadableStream<GetUserServiceLogsResponse>, Error>> = new Promise((resolve, _unusedReject) => {
+            const getUserServiceLogsStreamResponse: ClientReadableStream<GetUserServiceLogsResponse> = this.client.streamUserServiceLogs(getUserServiceLogsArgs);
+            resolve(ok(getUserServiceLogsStreamResponse));
+        })
+
+        const streamUserServiceLogsResponseResult: Result<ClientReadableStream<GetUserServiceLogsResponse>, Error> = await streamUserServiceLogsPromise;
+        if (streamUserServiceLogsResponseResult.isErr()) {
+            return err(streamUserServiceLogsResponseResult.error);
+        }
+
+        const streamUserServiceLogsResponse: ClientReadableStream<GetUserServiceLogsResponse> = streamUserServiceLogsResponseResult.value;
+
+        const userServiceLogsByGuid: Map<ServiceGUID, Array<string>> = new Map<ServiceGUID, Array<string>>();
+
+        const userServiceLogsReadable: Readable = this.createNewUserServiceLogsReadable(streamUserServiceLogsResponse);
+
+        streamUserServiceLogsResponse.on(GRPC_STREAM_RESPONSE_DATA_EVENT_NAME, function (getUserServiceLogsResponse: GetUserServiceLogsResponse) {
+
+            const userServiceLogsByUserServiceGuidMap: jspb.Map<string, LogLine> | undefined = getUserServiceLogsResponse.getUserServiceLogsByUserServiceGuidMap();
+
+            if (userServiceLogsByUserServiceGuidMap !== undefined) {
+                userServiceLogsByUserServiceGuidMap.forEach(
+                    (userServiceLogLine, userServiceGuidStr) => {
+                        userServiceLogsByGuid.set(userServiceGuidStr, userServiceLogLine.getLineList());
+                    }
+                )
+            }
+
+            userServiceLogsReadable.push(userServiceLogsByGuid);
+        })
+
+        streamUserServiceLogsResponse.on(GRPC_STREAM_RESPONSE_ERROR_EVENT_NAME, (streamLogsErr) => {
+            if (!userServiceLogsReadable.destroyed) {
+                //Propagate the GRPC error to the user service logs readable
+                const grpcStreamErr = new Error(`An error has been returned from the user service logs GRPC stream. Error:\n ${streamLogsErr}`)
+                userServiceLogsReadable.emit(USER_SERVICE_LOGS_READABLE_ERROR_EVENT_NAME, grpcStreamErr);
+            }
+        })
+
+        streamUserServiceLogsResponse.on(GRPC_STREAM_RESPONSE_END_EVENT_NAME, function () {
+            //Emit streams 'end' event when the GRPC stream has end
+            if (!userServiceLogsReadable.destroyed) {
+                userServiceLogsReadable.emit(USER_SERVICE_LOGS_READABLE_END_EVENT_NAME);
+            }
+        })
+
+        return ok(userServiceLogsReadable);
+    }
+
+    private createNewUserServiceLogsReadable(streamUserServiceLogsResponse: ClientReadableStream<GetUserServiceLogsResponse>): Readable {
+        const userServiceLogsReadable: Readable = new Readable({
+            objectMode: true, //setting object mode is to allow pass objects in the readable.push() method
+            read() {
+            } //this is mandatory to implement, we implement empty as it's describe in the implementation examples here: https://nodesource.com/blog/understanding-streams-in-nodejs/
+        })
+        userServiceLogsReadable.on(USER_SERVICE_LOGS_READABLE_CLOSE_EVENT_NAME, function () {
+            //Cancel the GRPC stream when when users close the ServiceLogsReadable
+            streamUserServiceLogsResponse.cancel();
+        })
+        return userServiceLogsReadable;
+    }
 }
+