+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/render_templates/render_templates.go
@@ -2,7 +2,6 @@ package render_templates
 
 import (
 	"context"
-	"fmt"
 	"github.com/kurtosis-tech/kurtosis/api/golang/core/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/service_network"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_instruction"
@@ -12,7 +11,6 @@ import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_validator"
 	"github.com/kurtosis-tech/stacktrace"
 	"go.starlark.net/starlark"
-	"strings"
 )
 
 const (
@@ -24,28 +22,39 @@ const (
 type RenderTemplatesInstruction struct {
 	serviceNetwork service_network.ServiceNetwork
 
-	position                          kurtosis_instruction.InstructionPosition
+	position       kurtosis_instruction.InstructionPosition
+	starlarkKwargs starlark.StringDict
+
 	templatesAndDataByDestRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData
 }
 
 func GenerateRenderTemplatesBuiltin(instructionsQueue *[]kurtosis_instruction.KurtosisInstruction, serviceNetwork service_network.ServiceNetwork) func(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
 	// TODO: Force returning an InterpretationError rather than a normal error
 	return func(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
-		templatesAndData, interpretationError := parseStartosisArgs(b, args, kwargs)
-		if interpretationError != nil {
+		renderTemplatesInstruction := newEmptyRenderTemplatesInstruction(serviceNetwork, *shared_helpers.GetCallerPositionFromThread(thread))
+
+		if interpretationError := renderTemplatesInstruction.parseStartosisArgs(b, args, kwargs); interpretationError != nil {
 			return nil, interpretationError
 		}
-		renderTemplatesInstruction := NewRenderTemplatesInstruction(serviceNetwork, *shared_helpers.GetCallerPositionFromThread(thread), templatesAndData)
 		*instructionsQueue = append(*instructionsQueue, renderTemplatesInstruction)
 		return starlark.String(renderTemplatesInstruction.position.MagicString(shared_helpers.ArtifactUUIDSuffix)), nil
 	}
 }
 
-func NewRenderTemplatesInstruction(serviceNetwork service_network.ServiceNetwork, position kurtosis_instruction.InstructionPosition, templatesAndDataByDestRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData) *RenderTemplatesInstruction {
+func newEmptyRenderTemplatesInstruction(serviceNetwork service_network.ServiceNetwork, position kurtosis_instruction.InstructionPosition) *RenderTemplatesInstruction {
+	return &RenderTemplatesInstruction{
+		serviceNetwork: serviceNetwork,
+		position:       position,
+		starlarkKwargs: starlark.StringDict{},
+	}
+}
+
+func NewRenderTemplatesInstruction(serviceNetwork service_network.ServiceNetwork, position kurtosis_instruction.InstructionPosition, templatesAndDataByDestRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData, starlarkKwargs starlark.StringDict) *RenderTemplatesInstruction {
 	return &RenderTemplatesInstruction{
 		serviceNetwork:                    serviceNetwork,
 		position:                          position,
 		templatesAndDataByDestRelFilepath: templatesAndDataByDestRelFilepath,
+		starlarkKwargs:                    starlarkKwargs,
 	}
 }
 
@@ -54,24 +63,7 @@ func (instruction *RenderTemplatesInstruction) GetPositionInOriginalScript() *ku
 }
 
 func (instruction *RenderTemplatesInstruction) GetCanonicalInstruction() string {
-	buffer := new(strings.Builder)
-	buffer.WriteString(RenderTemplatesBuiltinName + "(")
-	buffer.WriteString(templateAndDataByDestinationRelFilepathArg + "=")
-	numberOfTemplates := len(instruction.templatesAndDataByDestRelFilepath)
-	index := 0
-	for destinationPath, templateData := range instruction.templatesAndDataByDestRelFilepath {
-		index += 1
-		buffer.WriteString(fmt.Sprintf("\"%v\":{", destinationPath))
-		buffer.WriteString(fmt.Sprintf("\"template\":\"%v\"", templateData.Template))
-		buffer.WriteString(", ")
-		buffer.WriteString(fmt.Sprintf("\"template_data_json\":\"%v\"", templateData.DataAsJson))
-		buffer.WriteString("}")
-		if index < numberOfTemplates {
-			buffer.WriteString(", ")
-		}
-	}
-	buffer.WriteString(")")
-	return buffer.String()
+	return shared_helpers.CanonicalizeInstruction(RenderTemplatesBuiltinName, instruction.starlarkKwargs, &instruction.position)
 }
 
 func (instruction *RenderTemplatesInstruction) Execute(ctx context.Context, environment *startosis_executor.ExecutionEnvironment) error {
@@ -93,17 +85,17 @@ func (instruction *RenderTemplatesInstruction) ValidateAndUpdateEnvironment(envi
 	return nil
 }
 
-func parseStartosisArgs(b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData, *startosis_errors.InterpretationError) {
-
+func (instruction *RenderTemplatesInstruction) parseStartosisArgs(b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) *startosis_errors.InterpretationError {
 	var templatesAndDataArg *starlark.Dict
 	if err := starlark.UnpackArgs(b.Name(), args, kwargs, templateAndDataByDestinationRelFilepathArg, &templatesAndDataArg); err != nil {
-		return nil, startosis_errors.NewInterpretationError(err.Error())
+		return startosis_errors.NewInterpretationError(err.Error())
 	}
+	instruction.starlarkKwargs[templateAndDataByDestinationRelFilepathArg] = templatesAndDataArg
 
 	templatesAndDataByDestRelFilepath, interpretationErr := kurtosis_instruction.ParseTemplatesAndData(templatesAndDataArg)
 	if interpretationErr != nil {
-		return nil, interpretationErr
+		return interpretationErr
 	}
-
-	return templatesAndDataByDestRelFilepath, nil
+	instruction.templatesAndDataByDestRelFilepath = templatesAndDataByDestRelFilepath
+	return nil
 }